% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-skrmdb.r
\name{getdata}
\alias{getdata}
\title{Accessor to retrieve the the data used in the call to create an skrmdb object}
\usage{
getdata(x)
}
\arguments{
\item{x}{class skrmdb.}
}
\value{
a dataframe containing the sorted data.
}
\description{
This is an accessor function for retrieving the data used to compute the effective 
median dose from a \code{\link{skrmdb-class}} object object generated by \code{ReedMuench}, 
\code{SpearKarb}, or \code{DragBehr}.
}
