% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-skrmdb.r
\docType{methods}
\name{getED}
\alias{getED}
\alias{getED,skrmdb-method}
\title{Accessor to retrieve the numeric median effective dose from a skrmdb or sk object}
\usage{
getED(object)

\S4method{getED}{skrmdb}(object)
}
\arguments{
\item{object}{class skrmdb or sk}
}
\description{
This is an accessor function for retrieving the numeric value of effective 
from a skrmdb or sk data object object generated by ReedMuench, SpearKarb or DragBehr
}
\examples{
## with an object of class skrmdb
temp1 <- DragBehr(y=c(0,3,5,8,10,10), n=rep(10,6), x=1:6)
getED(temp1)
##  with an object of class sk
X <- data.frame(dead=c(0,3,5,8,10), total=rep(10,5), dil=1:5)
temp2 <- SpearKarb(cbind(dead,total) ~ dil, X)
getED(temp2)
}
\seealso{
\code{\link{skrmdb-class}}, \code{\link{sk-class}}
}
\author{
\link{skrmdb-package}
}
