% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-skrmdb.r
\name{getED}
\alias{getED}
\title{Accessor to retrieve the numeric median effective dose from a skrmdb
object}
\usage{
getED(x)
}
\arguments{
\item{x}{class skrmdb.}
}
\value{
the estimated ED50.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} This is an accessor function for
retrieving the numeric value of the effective median dose from a
\linkS4class{skrmdb} object object generated by \code{ReedMuench}, \code{SpearKarb}, or
\code{DragBehr}.
}
\keyword{internal}
