% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-skrmdb.r
\name{print.skrmdb}
\alias{print.skrmdb}
\title{print.skrmdb}
\usage{
\method{print}{skrmdb}(x, rnd = 3, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{skrmdb}.}

\item{rnd}{The number of digits to round the numeric values.}

\item{verbose}{If \code{TRUE}, prints out messages to help interpret the last five
columns of the printed table.}

\item{...}{Additional parameters.}
}
\value{
The \code{results} table in \code{x}, invisibly.
}
\description{
A print generic for a \verb{skrmdb object}
}
