% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skrmdb-package.r
\docType{package}
\name{skrmdb-package}
\alias{skrmdb-package}
\alias{skrmdb}
\title{skrmdb package.}
\description{
Includes functions to estimate ED50 by the methods of Spearman-Karber, Reed-Muench, Dragstedt-Behrens. \cr \cr For internal use only at
the USDA Center for Veterinary Biologics \cr
}
\details{
\tabular{ll}{
Package: \tab skrmdb-package\cr
Type: \tab Package\cr
Version: \tab 4.2.5\cr
Date: \tab 2018-08-03\cr
License: \tab MIT \cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\section{Resources}{


\itemize{
\item QUICK START: \url{https://github.com/ABS-dev/skrmdb/blob/master/README.md}
\item PACKAGE VIGNETTE: \url{https://www.aphis.usda.gov/animal_health/vet_biologics/publications/STATWI0001.pdf}
\item BUG REPORTS: \url{https://github.com/ABS-dev/skrmdb/issues}
}
}

\author{
David Siev \email{David.Siev@aphis.usda.gov}
}
