% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skrmdb.all.r
\name{skrmdb.all}
\alias{skrmdb.all}
\title{Function to run all three methods for determining ED50.}
\usage{
skrmdb.all(x, ..., autosort = FALSE)
}
\arguments{
\item{x}{A data frame containing the required columns \code{positive}, \code{total}, and 
\code{log_dil}}

\item{...}{The columns which are used to subset the data when calculating ED50 by the three 
methods.  If \code{...} is missing, the code will subset based on all columns that are not
the required ones.  If \code{... == NULL}, the code does not subset the data.}

\item{autosort}{Default \code{TRUE}.  If \code{TRUE} will sort the data according to either
\code{sort(x)} or \code{sort(-x)} so that \code{y / n} appears to be increasing with the
index.  This is how the three methods assume the data to be ordered.}
}
\value{
A \code{\link{data.frame}} containing columns for each of the subset variables, the ED50
  as computed by \code{DragBehr}, \code{ReedMuench}, the ED50 and variance as computed by 
  \code{SpearKarb}, and columns of logical values which indicate if the data are increasing or
  decreasing with \code{log_dil}, has an even dilution scheme, are monotone, and brackets the
  midpoint.
}
\description{
This function is used primarily by the CVB statistics section to run all three methods included
in the \code{\link{skrmdb}} package for determining ED50.
}
\examples{
data(titration)
titration$log_dil = -log10(titration$dil)
skrmdb.all(titration, Vial, Operator)

titration$dil = NULL
skrmdb.all(titration)
  
}
