% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragbehr.r, R/reedmuench.r, R/skrmdb.all.r,
%   R/skrmdb.r, R/spearkarb.r
\name{DragBehr}
\alias{DragBehr}
\alias{ReedMuench}
\alias{skrmdb.all}
\alias{skrmdb}
\alias{SpearKarb}
\title{Functions for computing ED50}
\usage{
DragBehr(
  data,
  formula,
  autosort = TRUE,
  warn.me = TRUE,
  show = FALSE,
  y = deprecated(),
  n = deprecated(),
  x = deprecated()
)

ReedMuench(
  data,
  formula,
  autosort = TRUE,
  warn.me = TRUE,
  show = FALSE,
  y = deprecated(),
  n = deprecated(),
  x = deprecated()
)

skrmdb.all(data, formula, autosort = TRUE, warn.me = TRUE)

SpearKarb(
  data,
  formula,
  autosort = TRUE,
  warn.me = TRUE,
  show = FALSE,
  y = deprecated(),
  n = deprecated(),
  x = deprecated()
)
}
\arguments{
\item{data}{A \code{data.frame} containing the titration data. Formatted as
specified in the CVB Data Guide.}

\item{formula}{A formula of the form \code{y + n ~ x} or \code{y + n ~ x | w1 + ... + wn} where \code{w1} ... \code{wn} are grouping variables.  All variables must be
distinct.}

\item{autosort}{If \code{TRUE}, the functions will compute the ED50 based on
either \code{y / n} or \code{1 - y / n}, whichever appears to increase with \code{x}. This
is how the three methods assume the data to be ordered. If \code{FALSE}, ED50
will be computed using \code{y / n}, which could give incorrect results.  Do not
change this parameter unless you are certain you know what you are doing.}

\item{warn.me}{If \code{TRUE}, warnings and messages related to the processing of
the data will be displayed.  These warnings correspond to the assumptions
and feedback columns in the output as outlined in the \strong{Description}
section above.}

\item{show}{If \code{TRUE}, will print the intermediate statistics used to
calculate ED50.  These statistics are captured in the data component of the
output.}

\item{y}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} An integer vector corresponding to the
number responding at each log dilution or dose.}

\item{n}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} An integer vector corresponding to the group
size at each log dilution or dose.}

\item{x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} A vector corresponding to the log dilution
or dose for each group.}
}
\value{
A list of class \link[=skrmdb-class]{skrmdb} which contains the following elements
\itemize{
\item \code{eval}: Which method or methods were used to compute ED50.
\item \code{data}: The transformed data used to compute ED50. It contains the
columns \code{y}, \code{n}, \code{x}, and any extra columns on which the results are
conditioned.  The columns \code{y_inc} and \code{y_dec} contain the increasing and
decreasing versions of the response proportions, based on \code{y / n} and \code{1 -   y / n}.  The columns \code{Duplicate}, \code{Dilutions}, \code{Response}, \code{Monotonic}, and
\code{Bracket} give information about the how well the data for each group meets
each of the assumptions and can be used to find specific parts of the data
which do not meet a particular assumption.
\item \code{ed}:  ED50.  If returned from \code{skrmdb.all()}, then \code{NA}.
\item \code{var}: The variance.  Only meaningful for \code{SpearKarb()}.
\item \code{results}: A \code{data.frame} giving the ED50 for each group based on the
conditions.  it also contains the columns \code{Duplicate}, \code{Dilutions},
\code{Response}, \code{Monotonic}, and \code{Bracket} which describe how well the data for
that group meets each of the assumptions.
\item \code{autosort}: The value of the parameter \code{autosort}.
}
}
\description{
The skrmdb package provides functionality to compute the median
effective dose (ED50) using the Dragstedt-Behrens, Reed-Muench, and
Spearman-Kärber estimators.

The Dragstedt-Behrens and Reed-Muench methods estimate the median effective
dose by interpolating between the two doses that bracket the dose producing
median response. They accumulate sums in both directions by assuming that
subjects that responded at a lower dose would respond at a higher dose, and
subjects that did not respond at a higher dose would not respond at a lower
dose. The Dragstedt-Behrens method estimates ED50 by interpolating on the
line that connects the hypothetical fractions of the bracketing doses for
ED50, while the Reed-Muench method estimates ED50 as the intersection of
the lines connecting the two sets of cumulative sums between bracketing
doses.

The Spearman-Kärber method gives a non-parametric estimate of the mean of a
tolerance distribution from its empirical distribution (EDF). The empirical
PMF is derived from the EDF by differencing.  The estimator is \eqn{\sum{ x
  f(x)}}{\sum x f(x)}. If the EDF does not cover the entire support of \code{x},
\code{SpearKarb()} extends it by assuming the next lower dilution would produce
zero response and the next higher dilution would produce complete response.

The function \code{skrmdb.all()} reports the results for all three methods.

There are several assumptions that these methods make:
\enumerate{
\item The group size \eqn{n_i} is constant.
\itemize{
\item Group sizes and responses are scaled silently if this condition is
not met.
}
\item The log dilutions \eqn{\mathbf{x} = \{x_1, x_2, \dots, x_k\}} form an
increasing sequence.
\itemize{
\item The data will be automatically sorted so this condition holds.  If
a dilution \eqn{x} appears more than once in \eqn{\mathbf{x}}, a new
dilution sequence will be created where each dilution appears once.  For
each dilution \eqn{x}, the corresponding \eqn{n_i} values are summed to
create a new \eqn{n} sequence, and the corresponding \eqn{y_i} values are
summed to create a new \eqn{y} sequence.  The column \strong{Duplicate} in the
output reports "Yes" if a dilution appears more than once in the original
sequence and "No" otherwise.
}
\item That \eqn{\mathbf{x}} is an arithmetic sequence. That is, there is some
number \eqn{\Delta} such that \eqn{x_k = x_1 + (k - 1) \times \Delta} for
all \eqn{k}.
\itemize{
\item The column \strong{Dilutions} in the output reports "regular" if this
condition is met and "irregular" otherwise.  Generally, this is used to
indicate the possibility of a missing dilution in a sequence.  However, it
is possible that in some cases, a sequence which is marked as "irregular"
may in fact be regular and have no missing dilutions.  In these cases, use
your best judgment.
}
\item The count ratio sequence \eqn{\mathbf{r} = \{r_1, r_2, \dots, r_k\}}
defined by \eqn{r_i = y_i / n_i} trends from smaller to larger.
\itemize{
\item The column \strong{Response} indicates whether \eqn{\mathbf{r}} is
"increasing", "decreasing", or shows "no change".  If it is "decreasing"
and \code{autosort} is \code{TRUE}, then ED50 will be computed using the sequence
\eqn{1 - r_i}, which is the appropriate transformation for decreasing
response data.  When the data is very noisy and the trend is not clear, it
is likely that using these methods to compute ED50 is not appropriate.
}
\item Preferably, the count ratio sequence is monotonic, that is \eqn{r_i \leq
  r_{i+1}} for all \eqn{i} or \eqn{r_i \geq r_{i+1}} for all \eqn{i}.
\itemize{
\item The column \strong{Monotonic} indicates if this condition holds.
Real-world data is rarely monotonic, but good data is generally "close
enough to monotonic for all practical purposes."  Once again, use your best
judgment.
}
\item The log dilution sequence brackets ED50.  That is, there is an \eqn{i}
such that \eqn{x_i \leq \text{ED50} \leq x_{i+1}}.  Note: This is \strong{not}
equivalent to saying that there is an \eqn{i} such that \eqn{r_i \leq 0.5
  \leq r_{i + 1}}.  However, if it is the case that \eqn{0.5 < r_1} or
\eqn{r_k < 0.5}, then the log dilution sequence will not bracket ED50.
\itemize{
\item The column \strong{Bracket} indicates if this condition holds.  If it
does not hold, use your best judgment as to whether or not ED50 could be
considered to be meaningful.
}
}
}
\note{
Many microbiology texts mistakenly present the Dragstedt-Behrens method
as the Reed-Muench method.
}
\examples{
# Processing Data with grouping variables.

titration$log_dil <- -log10(titration$dil)

SpearKarb(titration, positive + total ~ log_dil)
SpearKarb(titration, positive + total ~ log_dil | Operator)
SpearKarb(titration, positive + total ~ log_dil | Vial)
SpearKarb(titration, positive + total ~ log_dil | Operator + Vial)

skrmdb.all(titration, positive + total ~ log_dil | Operator + Vial)

## Monotonically increasing data

# The three calls are equivalent.
dead  <- c(0, 3, 5, 8, 10, 10)
total <- rep(10, 6)
dil   <- 1:6

## Use numeric vectors in the formula
DragBehr(dead + total ~ dil)
ReedMuench(dead + total ~ dil)
SpearKarb(dead + total ~ dil)
skrmdb.all(dead + total ~ dil)

data  <- data.frame(y = dead,
                    n = total,
                    x = dil)

## Use data plus the formula
DragBehr(data, y + n ~ x)
ReedMuench(data, y + n ~ x)
SpearKarb(data, y + n ~ x)
skrmdb.all(data, y + n ~ x)


## Using y, n, x is deprecated.
DragBehr(y = dead, n = total, x = dil) |> suppressWarnings()
ReedMuench(y = dead, n = total, x = dil) |> suppressWarnings()
SpearKarb(y = dead, n = total, x = dil) |> suppressWarnings()

# The function automatically automatically reorders the data as needed to
# meet the ordering assumption.
dead  <- rev(dead)
total <- rev(total)
dil   <- rev(dil)
SpearKarb(dead + total ~ dil)

## Unordered data
# Observe that the data is not monotonic after being sorted by dil.
dead  <- c(10, 8, 5, 3, 0)
total <- rep(10, 5)
dil   <- c(1, 3, 2, 4, 5)
SpearKarb(dead + total ~ dil)
}
\references{
Behrens, B. (1929) Zur Auswertung der Digitalisblätter im
Froschversuch. \emph{Arkiv für Experimentelle Pathologie und Pharmakologie.}
\strong{140: 237-256}.

Dragstedt, C. A., Lang, V. F. (1928). Respiratory Stimulants in Acute
Cocaine Poisoning in Rabbits. \emph{J. of Pharmacology and Experimental
Therapeutics.} \strong{32: 215--222}.

Kärber, G. (1931). Beitrag zur kollektiven Behandlung Pharmakologischer
Reihenversuche. \emph{Archiv für Experimentelle Pathologie und Pharmakologie.}
\strong{162: 480--483}.

Miller, Rupert G. (1973). Nonparametric Estimators of the Mean Tolerance in
Bioassay. \emph{Biometrika.} \strong{60: 535 - 542}.

Reed, L.J., Muench, H. (1938). A Simple Method of Estimating Fifty Percent
Endpoints. \emph{American Journal of Hygiene.} \strong{27: 493--497}.

Spearman, C. (1908). The Method of "Right and Wrong Cases" ("Constant
Stimuli") without Gauss's Formulae. \emph{Brit. J. of Psychology.} \strong{2:
227--242}.
}
