% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-skrmdb.r
\name{getED}
\alias{getED}
\title{Accessor to retrieve the numeric median effective dose from a skrmdb object}
\usage{
getED(x)
}
\arguments{
\item{x}{class skrmdb.}
}
\value{
the estimated ED50.
}
\description{
This is an accessor function for retrieving the numeric value of the
effective median dose from a \code{\link{skrmdb-class}} object object
generated by \code{ReedMuench}, \code{SpearKarb}, or \code{DragBehr}.
}
\examples{
temp <- DragBehr(y=c(0, 3,5, 8, 10, 10), n=rep(10, 6), x=1:6)
getED(temp)
}
