% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-skrmdb.r
\docType{class}
\name{skrmdb-class}
\alias{skrmdb-class}
\alias{skrmdb}
\title{Class definition for skrmdb object}
\description{
The skrmdb object holds output from functions in the skrmdb package.
}
\details{
\describe{
		\item{\code{eval}}{Evaluation method. One of 'ReedMuench', 'SpearKarb',  or 'DragBehr'. Character string.}
		\item{\code{ed}}{Median effective dose by eval method. Numeric.}
}
}

\examples{
new("skrmdb", ed = 2.906593, eval = "DragBehr")
}
\seealso{
\code{\link{sk-class}}
}
\author{
\link{skrmdb-package}
}
