\docType{package}
\name{skrmdb-package}
\alias{skrmdb-package}
\title{skrmdb package.}
\description{
Includes functions to estimate ED50 by the methods of
Spearman-Karber, Reed-Muench, Dragstedt-Behrens. \cr \cr
For internal use only at the USDA Center for Veterinary
Biologics \cr
}
\details{
\tabular{ll}{ Package: \tab skrmdb-package\cr Type: \tab
Package\cr Version: \tab 4.2.4\cr Date: \tab 2014-01-28\cr
License: \tab MIT \cr LazyLoad: \tab yes\cr LazyData: \tab
yes\cr }
}
\author{
David Siev \email{David.Siev@aphis.usda.gov}
}

