% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skrmdb.R
\docType{package}
\name{skrmdb-package}
\alias{skrmdb}
\alias{skrmdb-package}
\title{skrmdb: Package to estimate ED50}
\description{
Package to estimate ED50 by the methods of Spearman-Karber,
    Reed-Muench, Dragstedt-Behrens. No endorsement, claim, or warranty is 
    implied for this package. It is made available for investigational or pedagogical use only.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ABS-dev/skrmdb/blob/master/README.md}
  \item Report bugs at \url{https://github.com/ABS-dev/skrmdb/issues}
}

}
\author{
\strong{Maintainer}: CVB Statistics \email{CVB.Data.Help@usda.gov}

Authors:
\itemize{
  \item David Siev (until 2019) [former owner]
  \item CVB Statistics \email{CVB.Data.Help@usda.gov} (starting 2010)
}

}
\keyword{internal}
