% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-titration.r
\docType{data}
\name{titration}
\alias{titration}
\title{Example titration data.}
\format{
A data.frame with 67 rows and 9 columns
\itemize{
\item \code{testID}: Mandatory. A test identifier that is unique in the table.
Every test must have a test identifier.
\item \code{PrepID}: Mandatory. The identifier of the preparation used. This
will usually be a lot or serial number of a vaccine.
\item \code{PrepRole}: Mandatory. The role of the preparation. This must be
reference, test, or other.
\item \code{Date}: Optional. The Date the test was performed.
\item \code{Vial}: Optional. The vial number tested.
\item \code{Operator}: Optional. The operator who performed the test.
\item \code{dil}: Mandatory. The dilution used in a well.
\item \code{positive}: Mandatory. Total number of positive readings (tubes or
wells) affected by challenge.
\item \code{total}: Mandatory. The total number of tubes or wells in a group
(treatment group or at a specified dilution).
}
}
\usage{
titration
}
\description{
A dataset that illustrates the dichotomous data format for titration data as
found in the CVB DATA Guide.  The data is purposely incomplete and not-
monotonic.
}
\keyword{datasets}
